Move ActionTypes to a private export
